+++ src/Camera/style.ts
